requirements.txt
fastapi[standard]
uvicorn
transformers
torch
pillow